refactor(web): shared Base UI TimezoneCombobox; drop hand-rolled combo-box
The automation trigger timezone field now uses the same vendored
ui/combobox pattern as Business Info; the one-consumer ui/combo-box
is deleted.
